+++ pybit/__init__.py
@@ -1,1 +1,1 @@
-VERSION = "5.13.0"
+VERSION = "5.14.0rc0"