+++ pybit/market.py
@@ -18,6 +18,7 @@ class Market(str, Enum):
     GET_RISK_LIMIT = "/v5/market/risk-limit"
     GET_OPTION_DELIVERY_PRICE = "/v5/market/delivery-price"
     GET_LONG_SHORT_RATIO = "/v5/market/account-ratio"
+    GET_PRICE_LIMIT = "/v5/market/price-limit"
 
     def __str__(self) -> str:
         return self.value